feat: change reviews type Int to Long

+++ src/main/kotlin/org/appjam/smashing/domain/user/repository/UserSportProfileRepositoryCustomImpl.kt
@@ -124,7 +124,7 @@ class UserSportProfileRepositoryCustomImpl(
             )
 
         val reviewCountOrderExpression = Expressions.numberTemplate(
-            Int::class.java,
+            Long::class.java,
             "({0})",
             reviewCountExpression
         )
